pyproject.toml
[project]
name = "sudoku-trainer"
version = "0.1.0"
description = "Sudoku trainer with technique-aware solver and conversational tutor"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "anthropic>=0.49.0",
    "openai>=2.29.0",
    "google-genai>=1.68.0",
]
